@@ -15,55 +15,56 @@ dependencies:
   meta: any
 
   # Localization
-  intl: 0.18.1
+  intl: ^0.19.0
   flutter_localizations:
     sdk: flutter
 
   # State Management
-  flutter_riverpod: ^2.4.10
+  flutter_riverpod: ^2.5.1
 
   # Codegen
-  riverpod_annotation: ^2.3.4
+  riverpod_annotation: ^2.3.5
   freezed_annotation: ^2.4.1
-  json_annotation: ^4.8.1
+  json_annotation: ^4.9.0
 
   # Networking
-  dio: ^5.4.1
+  dio: ^5.4.3+1
   fresh_dio: ^0.4.1
   pretty_dio_logger: ^1.3.1
 
   # Persisting
-  shared_preferences: ^2.2.2
+  flutter_keychain: ^2.4.0
+  shared_preferences: ^2.2.3
   cached_network_image: ^3.3.1
 
   # Auth
-  amplify_flutter: ^1.7.0
-  amplify_auth_cognito: ^1.7.0
-  amplify_authenticator: ^1.5.4
-  firebase_core: ^2.24.2
-  firebase_auth: ^4.16.0
-  firebase_ui_auth: ^1.12.1
-  google_sign_in: ^6.2.1
-  firebase_ui_oauth_google: ^1.2.17
+  #  amplify_flutter: ^2.0.0
+  #  amplify_auth_cognito: ^2.0.0
+  #  amplify_authenticator: ^2.0.0
+  #  firebase_core: ^2.31.1
+  #  firebase_auth: ^4.20.0
+  #  firebase_ui_auth: ^1.14.0
+  #  google_sign_in: ^6.2.1
+  #  firebase_ui_oauth_google: ^1.3.2
 
   # Utils
-  go_router: ^13.2.0
-  path_provider: ^2.1.2
-  awesome_extensions: ^2.0.13
+  go_router: ^14.1.3
+  path_provider: ^2.1.3
+  awesome_extensions: ^2.0.16
   easy_debounce: ^2.0.3
-  share_plus: ^7.2.2
-  url_launcher: ^6.2.5
-  package_info_plus: ^5.0.1
+  share_plus: ^9.0.0
+  url_launcher: ^6.2.6
+  package_info_plus: ^8.0.0
 
   # UI
   flutter_image_slideshow: ^0.1.6
   flutter_rating_bar: ^4.0.1
-  google_fonts: ^6.1.0
+  google_fonts: ^6.2.1
   tree_select: ^0.0.1
-  readmore: ^2.2.0
+  readmore: ^3.0.0
   endless: ^1.6.3
-  chewie: ^1.7.5
-  video_player: ^2.8.3
+  chewie: ^1.8.1
+  video_player: ^2.8.6
   ms_undraw:
     git:
       url: https://github.com/MahMoos/ms_undraw.git
@@ -75,15 +76,15 @@ dev_dependencies:
   mocktail: ^1.0.3
 
   # Lints
-  custom_lint: ^0.6.2
-  riverpod_lint: ^2.3.9
+  custom_lint: ^0.6.4
+  riverpod_lint: ^2.3.10
   very_good_analysis: ^5.1.0
 
   # Codegen
-  freezed: ^2.4.7
-  build_runner: ^2.4.8
-  json_serializable: ^6.7.1
-  riverpod_generator: ^2.3.11
+  freezed: ^2.5.2
+  build_runner: ^2.4.10
+  json_serializable: ^6.8.0
+  riverpod_generator: ^2.4.0
 
   # Utils
   flutter_launcher_icons: ^0.13.1
