@@ -1,20 +1,20 @@
+import 'dart:async';
 import 'dart:core';
 
-import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_flutter/amplify_flutter.dart';
-import 'package:firebase_auth/firebase_auth.dart' as firebase;
+import 'package:flutter_keychain/flutter_keychain.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
+import '../../../common/api/api.dart';
 import '../../models/models.dart';
+import '../http/http_service.dart';
 
-part 'amplify_auth_service.dart';
+part 'agate_auth_service.dart';
 part 'auth_service.g.dart';
-part 'firebase_auth_service.dart';
+part 'mock_auth_service.dart';
 
 @Riverpod(keepAlive: true)
 Future<AuthService> authService(AuthServiceRef ref) async {
-  final service = AmplifyAuthService();
-  await service.initialize();
+  final service = AgateAuthService();
   return service;
 }
 
@@ -26,6 +26,12 @@ abstract class AuthService {
   String? get refreshToken;
 
   Future<String?> renewToken();
+
+  Stream<User?> userChanges();
+
+  Future<User> register(User user);
+
+  Future<User?> signInWithCredentials(String username, String password);
 
   Future<User?> signInWithProvider(AuthProviders provider);
 
